+++ packages/zowe-explorer/__tests__/__unit__/uss/actions.unit.test.ts
@@ -595,13 +595,11 @@ describe("USS Action Unit Tests - Functions uploadDialog & uploadFile", () => {
         Object.defineProperty(vscode.window, "withProgress", {
             value: jest.fn().mockImplementation((progLocation, callback) => {
                 const progress = {
-                    report: (message) => {
-                        return;
-                    },
+                    report: jest.fn(),
                 };
                 const token = {
                     isCancellationRequested: false,
-                    onCancellationRequested: undefined,
+                    onCancellationRequested: jest.fn(),
                 };
                 return callback(progress, token);
             }),